+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Fixed
 
-* Fixed an issue where Markdown headers stripped inline formatting (e.g. `## `code``).
+* Fixed an issue where Markdown headers stripped inline formatting (e.g. "## `code`").
 
 ## 1.3.4